attribute_example/applications/python/requirements.txt
fastapi
uvicorn[standard]
opentelemetry-api
opentelemetry-sdk
opentelemetry.instrumentation.fastapi
opentelemetry-exporter-otlp-proto-grpc
